+++ malachite-base/tests/strings/exhaustive/exhaustive_fixed_length_strings.rs
@@ -1,6 +1,6 @@
 use malachite_base::strings::exhaustive::exhaustive_fixed_length_strings;
 
-fn exhaustive_fixed_length_strings_helper(len: usize, out: &[&str]) {
+fn exhaustive_fixed_length_strings_helper(len: u64, out: &[&str]) {
     let css = exhaustive_fixed_length_strings(len)
         .take(20)
         .collect::<Vec<_>>();
